Conflicts:
	htdocs/langs/en_US/website.lang

+++ htdocs/core/class/commonobject.class.php
@@ -6537,7 +6537,7 @@ abstract class CommonObject
 			}
 		}
 
-		if (! $error)
+		if (! $error && ! empty($this->isextrafieldmanaged))
 		{
 			$sql = "DELETE FROM " . MAIN_DB_PREFIX . $this->table_element."_extrafields";
 			$sql.= " WHERE fk_object=" . $this->id;
